♻️ refactor: extract store
