+++ src/app/(with-header-sidebar)/mypage/_components/AlertModal.tsx
@@ -1,7 +1,7 @@
 import { ReactNode } from 'react';
 import Button from '@/components/Button';
 import styles from './AlertModal.module.css';
-import useModalStore from '../_store/modalStore';
+import useModalStore from '@/store/modalStore';
 
 export default function AlertModal({ children }: { children: ReactNode }) {
   const { closeModal } = useModalStore();
